feat(frontend): open console settings in dialog
Extract LLM settings into reusable panel, add workspace settings
dialog with capabilities/LLM tabs, replace navigation links with
dialog-opening buttons in console header and composer.
