+++ frontend/src/app/(console)/layout.tsx
@@ -3,7 +3,9 @@
 import Link from 'next/link'
 import { ThemeToggle } from '@/components/theme-toggle'
 import { LanguageToggle } from '@/components/language-toggle'
+import { WorkspaceSettingsDialog } from '@/components/settings/workspace-settings-dialog'
 import { useI18n } from '@/lib/i18n'
+import { useStore } from '@/lib/stores/context'
 
 export default function ConsoleLayout({
   children,
@@ -11,6 +13,7 @@ export default function ConsoleLayout({
   children: React.ReactNode
 }) {
   const { t } = useI18n()
+  const openWorkspaceSettings = useStore((state) => state.openWorkspaceSettings)
 
   return (
     <div className="min-h-screen flex flex-col bg-[radial-gradient(circle_at_top,rgba(34,211,238,0.18),transparent_20%),radial-gradient(circle_at_80%_20%,rgba(249,115,22,0.1),transparent_20%),linear-gradient(180deg,rgba(248,250,252,0.98)_0%,rgba(241,245,249,0.95)_55%,rgba(226,232,240,0.92)_100%)] text-foreground xl:h-screen xl:overflow-hidden dark:bg-[radial-gradient(circle_at_top,rgba(34,211,238,0.15),transparent_22%),radial-gradient(circle_at_80%_20%,rgba(249,115,22,0.12),transparent_20%),linear-gradient(180deg,#020617_0%,#06101f_55%,#030712_100%)] dark:text-foreground">
@@ -32,18 +35,20 @@ export default function ConsoleLayout({
               <span className="px-2 text-xs uppercase tracking-[0.22em] text-muted-foreground">
                 {t('settingsNav')}
               </span>
-              <Link
-                href="/console/capabilities"
+              <button
+                type="button"
                 className="rounded-full border border-cyan-300/35 bg-cyan-300/10 px-4 py-2 text-sm text-cyan-800 transition hover:bg-cyan-300/20 dark:text-cyan-100"
+                onClick={() => openWorkspaceSettings('capabilities')}
               >
                 {t('capabilitySettingsNav')}
-              </Link>
-              <Link
-                href="/console/llm"
+              </button>
+              <button
+                type="button"
                 className="rounded-full border border-cyan-300/35 bg-cyan-300/10 px-4 py-2 text-sm text-cyan-800 transition hover:bg-cyan-300/20 dark:text-cyan-100"
+                onClick={() => openWorkspaceSettings('llm')}
               >
                 {t('llmSettingsNav')}
-              </Link>
+              </button>
               <Link
                 href="/console/database"
                 className="rounded-full border border-cyan-300/35 bg-cyan-300/10 px-4 py-2 text-sm text-cyan-800 transition hover:bg-cyan-300/20 dark:text-cyan-100"
@@ -65,6 +70,7 @@ export default function ConsoleLayout({
       <main className="w-full flex-1 min-h-0 px-4 py-4 sm:px-6 sm:py-6 xl:overflow-y-auto xl:px-8 2xl:px-10">
         {children}
       </main>
+      <WorkspaceSettingsDialog />
     </div>
   )
 }
