+++ frontend/src/lib/i18n.ts
@@ -141,6 +141,8 @@ export const messages = {
     dockResultPanel: 'Dock Results',
     usePopupResults: 'Use Popup Results',
     resultPanelDialogTitle: 'Analysis Results & Report',
+    workspaceSettingsDialogTitle: 'Workspace Settings',
+    workspaceSettingsDialogDesc: 'Adjust model and capability settings without leaving the active conversation.',
     historyCollapsedTitle: 'History collapsed',
     historyCollapsedBody: 'Expand the panel to browse previous conversations.',
     newConversation: 'New Conversation',
@@ -583,6 +585,8 @@ export const messages = {
     dockResultPanel: '固定结果面板',
     usePopupResults: '弹窗显示结果',
     resultPanelDialogTitle: '分析结果与报告',
+    workspaceSettingsDialogTitle: '工作区设置',
+    workspaceSettingsDialogDesc: '无需离开当前对话即可调整模型与能力设置。',
     historyCollapsedTitle: '历史已收起',
     historyCollapsedBody: '展开面板后可浏览过往对话。',
     newConversation: '新建对话',
